DYN-4276-Step8-NavigateToPackages
I refactored some functions so the function that use reflection can be reused.
I added several comments to the new methods created.
