+++ src/DynamoCoreWpf/UI/GuidedTour/GuidesManager.cs
@@ -50,6 +50,9 @@ namespace Dynamo.Wpf.UI.GuidedTour
         private const string mainGridName = "mainGrid";
         private const string libraryViewName = "Browser";
 
+        /// <summary>
+        /// This property will be used for getting the current guide being executed in other assemblies like LibraryViewExtensionMSWebBrowser
+        /// </summary>
         public Guide CurrentGuide 
         {
             get
