+++ src/DynamoCoreWpf/UI/GuidedTour/Step.cs
@@ -147,6 +147,9 @@ namespace Dynamo.Wpf.UI.GuidedTour
 
         internal string GuideName { get; set; }
 
+        /// <summary>
+        /// This will give access to the Popup instance so we can find UIElements on it.
+        /// </summary>
         internal Popup StepUIPopup 
         {
             get
@@ -561,9 +564,10 @@ namespace Dynamo.Wpf.UI.GuidedTour
                     StepGuideBackground.GuideHighlightRectangle.Stroke = brush;
                 }
             }
-            //In case the item to be highlighted is inside the LibraryView (WebBrowser component) then we need to access the WebBrowser instance and call a js function
+            //This case is for when the item to be highlighted is inside the LibraryView (WebBrowser component) 
             else if (HostPopupInfo.HighlightRectArea.UIElementTypeString.Equals(typeof(WebBrowser).Name))
             {
+                //We need to access the WebBrowser instance and call a js function to highlight the html div border of the item
                 HighlightLibraryItem(bVisible);
             }
             //If the UIElementTypeString is not a MenuItem and also not a DynamoView we need to find the Window in the OwnedWindows and search the element inside it
@@ -638,79 +642,26 @@ namespace Dynamo.Wpf.UI.GuidedTour
             }
         }
 
+        /// <summary>
+        /// This method will call the collapseExpandPackage javascript method with reflection, so the package expander in LibraryView will be clicked
+        /// </summary>
         internal void CollapseExpandPackage()
-        {
-            const string webBrowserString = "Browser";
-            const string invokeScriptFunction = "InvokeScript";
+        {       
             const string jsMethodName = "collapseExpandPackage";
             const string packageName = "Autodesk Sample";
-
-            //Try to find the grid that contains the LibraryView
-            var sidebarGrid = (MainWindow as Window).FindName(HostPopupInfo.HostUIElementString) as Grid;
-            if (sidebarGrid == null) return;
-
-            //We need to iterate every child in the grid due that we need to apply reflection to get the Type and find the LibraryView (a reference to LibraryViewExtensionMSWebBrowser cannot be added).
-            foreach (var child in sidebarGrid.Children)
-            {
-                Type type = child.GetType();
-                if (type.Name.Equals(HostPopupInfo.WindowName))
-                {
-                    var libraryView = child as UserControl;
-                    //get the WebBrowser instance inside the LibraryView
-                    var browser = libraryView.FindName(webBrowserString);
-                    if (browser == null) return;
-
-                    Type typeBrowser = browser.GetType();
-                    //Due that there are 2 methods with the same name "InvokeScript", then we need to get the one with 2 parameters
-                    MethodInfo methodInvokeScriptInfo = typeBrowser.GetMethods().Single(m => m.Name == invokeScriptFunction && m.GetParameters().Length == 2);
-                    //Parameters for the WebBrowser.InvokeScript() function
-                    object[] parametersInvokeScript = new object[] { packageName };
-                    //Parameters for the JS Function collapseExpandPackage
-                    object[] parametersJSFunction = new object[] { jsMethodName, parametersInvokeScript };
-                    //Invoke the JS method located in library.html
-                    var resultHTML = methodInvokeScriptInfo.Invoke(browser, parametersJSFunction);
-                    break;
-                }
-            }
+            object[] parametersInvokeScript = new object[] { jsMethodName, new object[] { packageName } };
+            Guide.ExecuteJSFunction(MainWindow, HostPopupInfo, parametersInvokeScript);
         }
 
         /// <summary>
-        /// This method will highlight a item (<div></div>) in the WebBrowser instance located inside the LibraryView
-        /// due that we cannot include a reference to LibraryViewExtensionMSWebBrowser then we need to use reflection to get the types
+        /// This method will execute a js function for highlighting a item (<div></div>) in the WebBrowser instance located inside the LibraryView using reflection
         /// </summary>
         /// <param name="visible">enable or disable the highlight in a specific Library item</param>
         internal void HighlightLibraryItem(bool visible)
         {
-            const string webBrowserString = "Browser";
-            const string invokeScriptFunction = "InvokeScript";
             const string jsMethodName = "highlightLibraryItem";
-
-            var sidebarGrid = (MainWindow as Window).FindName(HostPopupInfo.HighlightRectArea.UIElementGridContainer) as Grid;
-            if (sidebarGrid == null) return;
-
-            //We need to iterate every child in the Children collection due that we need to apply reflection to get the Type
-            foreach (var child in sidebarGrid.Children)
-            {
-                Type type = child.GetType();
-                if (type.Name.Equals(HostPopupInfo.HighlightRectArea.WindowName))
-                {
-                    var libraryView = child as UserControl;
-                    //get the WebBrowser instance inside the LibraryView
-                    var browser = libraryView.FindName(webBrowserString); 
-                    if (browser == null) return;
-
-                    Type typeBrowser = browser.GetType();
-                    //Due that there are 2 methods with the same name "InvokeScript", then we need to get the one with 2 parameters
-                    MethodInfo methodInvokeScriptInfo = typeBrowser.GetMethods().Single(m => m.Name == invokeScriptFunction && m.GetParameters().Length == 2);
-                    //Parameters for the WebBrowser.InvokeScript() function
-                    object[] parametersInvokeScript = new object[] { HostPopupInfo.HighlightRectArea.WindowElementNameString, visible };
-                    //Parameters for the JS Function highlightLibraryItem
-                    object[] parametersJSFunction = new object[] { jsMethodName, parametersInvokeScript };
-                    //Invoke the JS method located in library.html
-                    var resultHTML = methodInvokeScriptInfo.Invoke(browser, parametersJSFunction);
-                    break;
-                }
-            }
+            object[] parametersInvokeScript = new object[] { jsMethodName, new object[] { HostPopupInfo.HighlightRectArea.WindowElementNameString, visible } };
+            Guide.ExecuteJSFunction(MainWindow, HostPopupInfo, parametersInvokeScript);
         }
 
         /// <summary>
